refactor: auth 필터 로직 리팩터링
