+++ src/main/java/com/nexters/goalpanzi/common/filter/JwtAuthenticationFilter.java
@@ -1,13 +1,18 @@
 package com.nexters.goalpanzi.common.filter;
 
-import com.nexters.goalpanzi.common.jwt.JwtParser;
-import com.nexters.goalpanzi.common.jwt.JwtProvider;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nexters.goalpanzi.common.auth.jwt.JwtParser;
+import com.nexters.goalpanzi.common.auth.jwt.JwtProvider;
+import com.nexters.goalpanzi.exception.ErrorCode;
+import com.nexters.goalpanzi.exception.ErrorResponse;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -16,17 +21,18 @@ import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
+@Component
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
     private static final String TOKEN_REISSUE_URI = "/api/auth/token:reissue";
     private static final String CONTENT_TYPE = "application/json";
     private static final String ATTRIBUTE_NAME = "altKey";
-    private static final String UNAUTHORIZED_MESSAGE = "{\"error\": \"Unauthorized\"}";
 
     private static final List<String> whitelist = List.of("/api/auth/login");
 
     private final JwtProvider jwtProvider;
     private final JwtParser jwtParser;
+    private final ObjectMapper objectMapper;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
@@ -49,16 +55,20 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
                 filterChain.doFilter(request, response);
                 return;
             }
-
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            response.setContentType(CONTENT_TYPE);
-            response.getWriter().write(UNAUTHORIZED_MESSAGE);
+            respondWithUnauthorized(response);
         } catch (Exception e) {
             log.error(e.getMessage(), e);
-            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-            response.setContentType(CONTENT_TYPE);
-            response.getWriter().write(e.getMessage());
+            respondWithUnauthorized(response);
         }
+    }
+
+    private void respondWithUnauthorized(final HttpServletResponse response) throws IOException {
+        response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+        response.setContentType(CONTENT_TYPE);
+        response.setCharacterEncoding("UTF-8");
+        response.getWriter().write(objectMapper.writeValueAsString(
+                new ErrorResponse(HttpStatus.UNAUTHORIZED.value(), ErrorCode.INVALID_TOKEN.getMessage()))
+        );
     }
 
     private Boolean isTokenReissueRequest(String requestURI, String token) {
