+++ src/main/java/com/nexters/goalpanzi/config/WebFilterConfig.java
@@ -1,8 +1,8 @@
 package com.nexters.goalpanzi.config;
 
 import com.nexters.goalpanzi.common.filter.JwtAuthenticationFilter;
-import com.nexters.goalpanzi.common.jwt.JwtParser;
-import com.nexters.goalpanzi.common.jwt.JwtProvider;
+import com.nexters.goalpanzi.common.auth.jwt.JwtParser;
+import com.nexters.goalpanzi.common.auth.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
@@ -12,13 +12,12 @@ import org.springframework.context.annotation.Configuration;
 @RequiredArgsConstructor
 public class WebFilterConfig {
 
-    private final JwtProvider jwtProvider;
-    private final JwtParser jwtParser;
+    private final JwtAuthenticationFilter jwtAuthenticationFilter;
 
     @Bean
     public FilterRegistrationBean<JwtAuthenticationFilter> jwtFilter() {
         FilterRegistrationBean<JwtAuthenticationFilter> filterRegistrationBean = new FilterRegistrationBean<>();
-        filterRegistrationBean.setFilter(new JwtAuthenticationFilter(jwtProvider, jwtParser));
+        filterRegistrationBean.setFilter(jwtAuthenticationFilter);
         filterRegistrationBean.addUrlPatterns("/api/*");
         return filterRegistrationBean;
     }
