+++ src/test/java/com/nexters/goalpanzi/application/auth/apple/AppleClaimsValidatorTest.java
@@ -1,6 +1,6 @@
 package com.nexters.goalpanzi.application.auth.apple;
 
-import com.nexters.goalpanzi.common.util.Nonce;
+import com.nexters.goalpanzi.common.auth.Nonce;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import org.junit.jupiter.api.Test;
